packages/ml/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
xgboost>=2.0.0
scikit-learn>=1.4.0
shap>=0.44.0
pandas>=2.2.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
pydantic>=2.5.0
joblib>=1.3.0
